add distanceto

+++ NC24/Point.h
@@ -16,7 +16,7 @@ class Point {
         unsigned int getY() const;
         bool setX(unsigned int m_x);
         bool setY(unsigned int m_y);
-        unsigned int distanceTo(const Point &other);
+        float distanceTo(const Point &other);
         void operator=(const Point& point);
         friend bool operator==(Point m_x, Point m_y);
         friend ostream& operator<<(ostream& out, const Point& p);
